requirements.txt
asgiref==3.8.1 ; python_version >= "3.11" and python_version < "3.13"
black==24.8.0 ; python_version >= "3.11" and python_version < "3.13"
click==8.1.7 ; python_version >= "3.11" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "3.13" and (sys_platform == "win32" or platform_system == "Windows")
django-filter==24.3 ; python_version >= "3.11" and python_version < "3.13"
django==5.1.1 ; python_version >= "3.11" and python_version < "3.13"
djangorestframework==3.15.2 ; python_version >= "3.11" and python_version < "3.13"
flake8==7.1.1 ; python_version >= "3.11" and python_version < "3.13"
iniconfig==2.0.0 ; python_version >= "3.11" and python_version < "3.13"
isort==5.13.2 ; python_version >= "3.11" and python_version < "3.13"
libcst==1.1.0 ; python_version >= "3.11" and python_version < "3.13"
markdown==3.7 ; python_version >= "3.11" and python_version < "3.13"
mccabe==0.7.0 ; python_version >= "3.11" and python_version < "3.13"
mypy-extensions==1.0.0 ; python_version >= "3.11" and python_version < "3.13"
packaging==24.1 ; python_version >= "3.11" and python_version < "3.13"
pathspec==0.12.1 ; python_version >= "3.11" and python_version < "3.13"
platformdirs==4.3.2 ; python_version >= "3.11" and python_version < "3.13"
pluggy==1.5.0 ; python_version >= "3.11" and python_version < "3.13"
psycopg2-binary==2.9.9 ; python_version >= "3.11" and python_version < "3.13"
pycodestyle==2.12.1 ; python_version >= "3.11" and python_version < "3.13"
pyflakes==3.2.0 ; python_version >= "3.11" and python_version < "3.13"
pytest==8.3.3 ; python_version >= "3.11" and python_version < "3.13"
pyyaml==6.0.2 ; python_version >= "3.11" and python_version < "3.13"
sqlparse==0.5.1 ; python_version >= "3.11" and python_version < "3.13"
toml==0.10.2 ; python_version >= "3.11" and python_version < "3.13"
tornado==6.4.1 ; python_version >= "3.11" and python_version < "3.13"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "3.13"
typing-inspect==0.9.0 ; python_version >= "3.11" and python_version < "3.13"
tzdata==2024.1 ; python_version >= "3.11" and python_version < "3.13" and sys_platform == "win32"
unimport==1.2.1 ; python_version >= "3.11" and python_version < "3.13"
